+++ CHANGELOG.md
@@ -18,6 +18,7 @@ as well as to [Module version numbering](https://go.dev/doc/modules/version-numb
   is called when `Flow.Run` returns `CodeInvalidArgs`.
 - Rename `Flow.Register` to `Flow.Define`.
 - Change `Flow.RegisteredTask` to sealed interface `Flow.DefinedTask`.
+- Change the `Flow.DefaultTask` field to `Flow.SetDefault` and `Flow.Default` methods.
 
 ## [2.0.0-rc.1](https://github.com/goyek/goyek/compare/v1.1.0...v2.0.0-rc.1) - 2022-10-14